Api: Always extend with all available protocols
This helps both with documentation purposes as well as allowing methods
to call between protocols, even if not explicitly requested.

Protocols are still currently stored, as the Application Service
extensions check for protocol existance before executing, which should
probably be done in another way as it's the only user of the protocols
list.