Extensions: make the ignore_inspect code safer
Before this there's a risk of an infinite recursion if two objects using
ignore_inspect contain eachother outside of the ignored symbol list.

Might have to think of a better way to do this later on, one that
doesn't require scouring the callstack on every object inspection. Even
if that particular performance degradation risk is not too horrible